Update dictionary parsing for loop
Previously the code would not loop through the entire water quality dictionary. This commit fixes that problem.

Currently several node treatments and the link treatments that require the new pyswmm getters are failing on my local machine. I believe it is my machine, not a problem with the code. Will continue to debug.